+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # 0.5
 
 * Several substitutions:
-  * `Abort.INSTANCE` and `Resume.INSTANCE` became methods in `Restarts`.
+  * `Resume.INSTANCE` became a method in `Restarts`.
   * `Handler.Operations.use()` is now a restart, `UseValue`.
   * `Condition.onHandlerNotFound` no longer exists; now `Scope.signal` takes a _policy_ as a parameter. As a result, `Notice`s and `Warning`s are no longer needed.
+  * `Abort` became a method in `Handler.Operations`.
 * `Scope.signal` is now generic, and so is `Restart`.
 * `HandlerOps` provides handler bodies for common uses.
-* `Scope.notify` and `Scope.raise` cover common `Scope.signal` uses, with better ergonomics. `Scope.signal` now works better as a primitive operation, backing the others.
+* `Scope.notify` and `Scope.raise` cover common `Scope.signal` uses, with better ergonomics. `Scope.signal` works as a primitive operation, backing the others.
 
 # 0.4